tests: moved core tests to tests/core folder
