+++ tests/core/fixtures/files/pkg.json
