+++ tests/core/fixtures/files/setup.py
