+++ tests/core/fixtures/hardware/string_any.txt
