+++ tests/core/fixtures/hardware/string_some.txt
