+++ tests/core/fixtures/objects/activeinactive-mode.txt
