+++ tests/core/fixtures/objects/set.txt
