+++ tests/core/fixtures/objects/single-mode.txt
