+++ tests/core/fixtures/package_string.txt
