+++ tests/core/fixtures/package_string_empty.txt
