+++ tests/core/test_compression.py
