+++ tests/core/test_install_condition.py
