+++ tests/core/test_modes.py
