+++ tests/core/test_object.py
