+++ tests/core/test_options.py
